% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diabetesprevalence}
\alias{diabetesprevalence}
\title{5002 publications on diabetes prevalence}
\format{
A data frame with 5002 rows and 25 variables.
}
\usage{
data(diabetesprevalence)
}
\description{
A dataset containing 5002 publications exported from a Citavi project.
}
\keyword{datasets}
