% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_file.R
\name{example_file}
\alias{example_file}
\title{Get path to example file}
\usage{
example_file(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, all example files will be listed.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr
CitaviR comes bundled with some example files in its \code{inst/extdata}
directory. This function makes them easy to access.
}
\examples{
example_file()
example_file("3dupsin5refs.xlsx")
example_file("3dupsin5refs/3dupsin5refs.ctv6")
}
