% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_xlsx.R
\name{example_xlsx}
\alias{example_xlsx}
\title{Get path to example xlsx exported from Citavi}
\usage{
example_xlsx(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, all example files will be listed.}
}
\description{
CitaviR comes bundled with some example files in its `inst/extdata`
directory. This function makes them easy to access.
}
\examples{
example_xlsx()
example_xlsx("3dupsin5refs.xlsx")
}
