% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obvious_dups.R
\name{find_obvious_dups}
\alias{find_obvious_dups}
\title{Identify obvious duplicates based on title and year}
\usage{
find_obvious_dups(CitDat, dupInfoAfterID = TRUE, preferDupsWithPDF = TRUE)
}
\arguments{
\item{CitDat}{A dataframe holding the information of the xls/xlsx file
\href{https://www1.citavi.com/sub/manual6/en/index.html?exporting_to_excel.html}{created with Citavi via export to Excel}.
Thus, this dataframe may have been imported via \code{\link[CitaviR]{read_Citavi_xlsx}}. The following columns
\bold{must be present} in the dataframe: \code{ID}, \code{Title}, \code{Year}.}

\item{dupInfoAfterID}{If TRUE (default), the newly created columns
\code{clean_title}, \code{clean_title_id}, \code{has_obv_dup} and \code{obv_dup_id}
are moved right next to the \code{ID} column.}

\item{preferDupsWithPDF}{If TRUE (default), obvious duplicates are sorted by their info
in columns \code{has_attachment} and/or \code{Locations} (given they are present in the dataset).
After sorting, duplicates with the most occurences of \code{".pdf"} in \code{Locations} and a
\code{TRUE} in \code{has_attachment} are first and will thus be chosen as \code{dup_01}.}
}
\value{
A tibble containing three additional columns:
\code{clean_title_id}, \code{has_obv_dup} and \code{obv_dup_id}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
Currently this only works for files that were generated while Citavi
was set to "English" so that column names are "Short Title" etc.
}
\examples{
path <- example_xlsx("3dupsin5refs.xlsx")
read_Citavi_xlsx(path) \%>\%
   find_obvious_dups()

}
