% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-not_in.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Helper function: Negation/Opposite of \%in\%}
\usage{
a \%not_in\% b
}
\arguments{
\item{a}{vector or NULL: the values to be matched.}

\item{b}{vector or NULL: the values to be matched against.}
}
\value{
A logical vector, indicating if a match was not located for each element of \code{a}.
The values are TRUE or FALSE and never NA.
}
\description{
\code{ \%not_in\% <- function(a,b){ ! a \%in\% b } }
}
\examples{
1:5 \%not_in\% 3:5

}
\keyword{utility}
