% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_obvious_dups.R
\name{handle_obvious_dups}
\alias{handle_obvious_dups}
\title{Handle obvious duplicates}
\usage{
handle_obvious_dups(CitDat, colsToWriteInto = "refInfo", nameDupCat = NULL)
}
\arguments{
\item{CitDat}{A dataframe returned by \code{\link[CitaviR]{find_obvious_dups}}.}

\item{colsToWriteInto}{Can be "refInfo", "all" or a custom character vector with column names. (TO DO)}

\item{nameDupCat}{What should the duplicate category be named? (TO DO)}
}
\value{
A tibble where information was spread across obvious duplicates. (TO DO)
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Currently this only works for files that were generated while Citavi
was set to "English" so that column names are "Short Title" etc.
}
\examples{
path <- example_xlsx("3dupsin5refs.xlsx")
read_Citavi_xlsx(path) \%>\%
   find_obvious_dups() \%>\%
   handle_obvious_dups()

}
