% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Citavi_ctv6.R
\name{read_Citavi_ctv6}
\alias{read_Citavi_ctv6}
\title{Read table from Citavi database (via SQL)}
\usage{
read_Citavi_ctv6(path = NULL, CitDBTableName = "Reference")
}
\arguments{
\item{path}{Path to the local Citavi project file (.ctv6).}

\item{CitDBTableName}{Name of the table to be read from the connected Citavi database (via \code{DBI::dbReadTable()}).
Set to "Reference" by default. Shows all table names when set to NULL (via \code{DBI::dbListTables}).}
}
\value{
A tibble
}
\description{
Read table from Citavi database (via SQL)
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr
The underlying core functions are \code{DBI::dbConnect()} \code{RSQLite::SQLite()}, \code{DBI::dbReadTable()} and \code{DBI::dbListTables}.
}
\examples{
# example Citavi project
example_path <- example_file("3dupsin5refs/3dupsin5refs.ctv6")

# import reference (=default) table
CitDat <- read_Citavi_ctv6(example_path)
CitDat \%>\% dplyr::select(Title, Year, Abstract, DOI)

# show table names
read_Citavi_ctv6(example_path, CitDBTableName = NULL)

}
