% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_Citavi_xlsx.R
\name{write_Citavi_xlsx}
\alias{write_Citavi_xlsx}
\title{Export}
\usage{
write_Citavi_xlsx(CitDat, read_path = NULL)
}
\arguments{
\item{CitDat}{A tibble of processed Citavi data TO DO}

\item{read_path}{The path used for \code{\link[CitaviR]{read_Citavi_xlsx}}.}
}
\value{
nothing ?! TO DO
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
#'
}
\examples{
\dontrun{
your_path <- "data/yourCitaviExport.xlsx"
read_Citavi_xlsx(your_path) \%>\%
   find_obvious_duplicates \%>\%
   handle_obvious_duplcates \%>\%
   write_Citavi_xlsx(your_path)
}

}
