% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_xlsx.R
\name{example_xlsx}
\alias{example_xlsx}
\title{Get path to example xlsx exported from Citavi}
\usage{
example_xlsx(file = NULL)
}
\arguments{
\item{file}{Name of file. If \code{NULL}, all example files will be listed.}
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#stable'><img src='figures/lifecycle-stable.svg' alt='Stable lifecycle'></a>}}{\strong{Stable}}
CitaviR comes bundled with some example files in its \code{inst/extdata}
directory. This function makes them easy to access.
}
\examples{
example_xlsx()
example_xlsx("3dupsin5refs.xlsx")
}
