% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obvious_dups.R
\name{find_obvious_dups}
\alias{find_obvious_dups}
\title{Identify obvious duplicates based on title and year}
\usage{
find_obvious_dups(CitDat)
}
\arguments{
\item{CitDat}{A dataframe holding the information of the xls/xlsx file
\href{https://www1.citavi.com/sub/manual6/en/index.html?exporting_to_excel.html}{created with Citavi via export to Excel}.
Thus, this dataframe may have been imported via \code{\link[CitaviR]{read_Citavi_xlsx}}. The following columns
\bold{must be present} in the dataframe: \code{ID}, \code{Title}, \code{Year}.}
}
\value{
A tibble containing three additional columns:
\code{clean_title_id}, \code{has_obv_dup} and \code{obv_dup_id}.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Currently this only works for files that were generated while Citavi
was set to "German" so that column names are "Kurztitel" etc.
}
\examples{
\dontrun{
path <- example_xlsx("3dupsin5refs.xlsx") # use this package's example xlsx file
read_Citavi_xlsx(path) \%>\%
   find_obvious_dups()
}

}
