% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_obvious_dups.R
\name{find_obvious_dups}
\alias{find_obvious_dups}
\title{Identify obvious duplicates based on title and year}
\usage{
find_obvious_dups(CitDat, dupInfoAfterID = TRUE, preferDupsWithPDF = TRUE)
}
\arguments{
\item{CitDat}{A dataframe/tibble returned by \code{\link[CitaviR]{read_Citavi_xlsx}}.
The following columns \bold{must be present}: \code{ID}, \code{Title}, \code{Year}.}

\item{dupInfoAfterID}{If TRUE (default), the newly created columns
\code{clean_title}, \code{clean_title_id}, \code{has_obv_dup} and \code{obv_dup_id}
are moved right next to the \code{ID} column. Additionally, the \code{ID} column is
moved to the first position.}

\item{preferDupsWithPDF}{If TRUE (default), obvious duplicates are sorted by their info
in columns \code{has_attachment} and/or \code{Locations} (given they are present in the dataset).
After sorting, duplicates with the most occurences of \code{".pdf"} in \code{Locations} and a
\code{TRUE} in \code{has_attachment} are first and will thus be chosen as \code{dup_01}.}
}
\value{
A tibble containing four additional columns:
\code{clean_title}, \code{clean_title_id}, \code{has_obv_dup} and \code{obv_dup_id}.
}
\description{
Identify obvious duplicates based on title and year
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}} \cr
Currently this only works for files that were generated while Citavi
was set to "English" so that column names are "Short Title" etc.
}
\examples{
example_path <- example_file("3dupsin5refs/3dupsin5refs.ctv6")
read_Citavi_ctv6(example_path) \%>\%
   find_obvious_dups() \%>\%
   dplyr::select(clean_title:obv_dup_id)

}
